+++ changelog.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Changed text in table of positions if no positions are imported
+- Extend support for extension regarding its PHP-strict compatibility
 
 ## [4.2.7] - 16.04.2025