+++ changelog.md
@@ -7,6 +7,10 @@
 - Changed text in table of positions if no positions are imported
 - Extend support for extension regarding its PHP-strict compatibility
 
+### Fixed
+
+- Fixed usage of form filter if simple permalinks are used
+
 ## [4.2.7] - 16.04.2025
 
 ### Changed