+++ changelog.md
@@ -7,6 +7,8 @@
 - Changed text in table of positions if no positions are imported
 - Optimized Block Editor loading
 - Extend support for extension regarding its PHP-strict compatibility
+- Optimized build process regarding check against WordPress Coding Standards
+- GitHub action does not fail if it automatically fixes code issues
 
 ### Fixed